% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Choice_Set}
\alias{Choice_Set}
\title{Collects a choice set exposed to individuals.}
\usage{
Choice_Set(df, num_id, uniq_id, idvar)
}
\arguments{
\item{df}{a long format tibble.}

\item{num_id}{Number of unique Booking_IDs appearing in transaction data}

\item{uniq_id}{Unique Booking_ID in transaction data.}

\item{idvar}{Variable name representing customer id (Booking_ID).}
}
\value{
Returns a list containing the values required for calculation within the \code{\link{rmm_reshape}} function.
}
\description{
Collects a choice set exposed to individuals. Internal function that users should not call directly.
}
