% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coalgov}
\alias{coalgov}
\title{Coalition governments in comparative perspective (1944-2013)}
\format{
A data frame with 1286 rows and 14 variables:
\describe{
  \item{Col1}{description of Col1}
...
}
}
\usage{
data(coalgov)
}
\description{
The unit of analysis for this dataset is level 1, which is the party level in this case.
pid indicates the party identity
gid indicates the government identity
...
}
\examples{
data(coalgov)
}
\keyword{datasets}
