% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.char.2.factor.R
\name{conv.char.2.factor}
\alias{conv.char.2.factor}
\title{conv.char.2.factor}
\usage{
conv.char.2.factor(df)
}
\arguments{
\item{df}{dataframe to run conversion on}
}
\value{
dataframe with all character columns converted to factors.
}
\description{
Convert character columns to factor columns.
}

