% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv.factor.2.char.R
\name{conv.factor.2.char}
\alias{conv.factor.2.char}
\title{conv.factor.2.char}
\usage{
conv.factor.2.char(df)
}
\arguments{
\item{df}{dataframe to run conversion on}
}
\value{
dataframe with all factor columns converted to character.
}
\description{
Convert factor columns to character columns.
}

