% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpaces.R
\name{countSpaces}
\alias{countSpaces}
\title{countSpaces}
\usage{
countSpaces(s)
}
\arguments{
\item{s}{string to count spaces}
}
\value{
return the count of spaces
}
\description{
Count the number of spaces in a string
}

