% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESTF.R
\name{ESTF}
\alias{ESTF}
\title{Estimating model parameters.}
\usage{
ESTF(Alts_Same_Choice_Set, Choice_Set_List, Num_Prod, Num_Attr, Is_Purchase)
}
\arguments{
\item{Alts_Same_Choice_Set}{list, for internal calculation, a list of data exposed to the same choice set.}

\item{Choice_Set_List}{list, product code included for each choice set.}

\item{Num_Prod}{numeric, the number of all products in observed data.}

\item{Num_Attr}{numeric, the number of attributes in observed data.}

\item{Is_Purchase}{list, for internal calculation, list of customers who are
exposed to the same choice set who purchased the product. If purchased 1, otherwise 0.}
}
\value{
The estimated value of parameters except gamma(no-purchase utility).
}
\description{
Estimating model parameters except gamma(no-purchase utility). Internal function that users should not call directly.
}
\keyword{internal}
