% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ew.R
\name{ew}
\alias{ew}
\title{ew}
\usage{
ew(...)
}
\arguments{
\item{...}{pass operation to evaluate}
}
\value{
list of res withCallingHandlers, err tryCatch error, warn tryCatch waring
}
\description{
Capture results Errors and Warnings.
}

