% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.columns.with.pattern.R
\name{find.columns.with.pattern}
\alias{find.columns.with.pattern}
\title{find.columns.with.pattern}
\usage{
find.columns.with.pattern(data, pattern)
}
\arguments{
\item{data}{dataframe to search}

\item{pattern}{pattern to grep}
}
\value{
column names
}
\description{
Return column name when column contains pattern.
}

