% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{govsurv}
\alias{govsurv}
\title{Artificial dataset on coalition government survival}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21 rows and 9 columns.}
\usage{
data(govsurv)
}
\description{
The unit of analysis for this dataset is level 1, which is the party level in this case.
pid indicates the party identity
gid indicates the government identity
}
\examples{
data(govsurv)
}
\keyword{datasets}
