% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike.R
\name{loglike}
\alias{loglike}
\title{Calculating the log-likelihood function.}
\usage{
loglike(
  eta,
  Choice_Set_List,
  Choice_Set_Code,
  Base_Prod,
  Num_Prod,
  Num_Attr,
  Alts_Same_Choice_Set,
  Is_Purchase
)
}
\arguments{
\item{eta}{numeric vector, the model parameters to be estimated}

\item{Choice_Set_List}{list, product code included for each choice set.}

\item{Choice_Set_Code}{numeric, choice set code Number specified for calculating log-likelihood.}

\item{Base_Prod}{numeric, the baseline product number}

\item{Num_Prod}{numeric, the number of all products in observed data.}

\item{Num_Attr}{numeric, the number of attributes in observed data.}

\item{Alts_Same_Choice_Set}{list, for internal calculation, a list of data exposed to the same choice set.}

\item{Is_Purchase}{list, for internal calculation, list of customers who are
exposed to the same choice set who purchased the product. If purchased 1, otherwise 0.}
}
\value{
The value of log-likelihood.
}
\description{
Calculating model's log-likelihood function. Internal function that users should not call directly.
}
\keyword{internal}
