% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{mid}
\alias{mid}
\title{mid}
\usage{
mid(text, start_num, num_char)
}
\arguments{
\item{text}{string}

\item{start_num}{number of chars from the left to strat at}

\item{num_char}{number of chars from the start point}
}
\value{
left substring
}
\description{
substring ofstring.
}

