% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monetPlot.R
\name{monetPlot}
\alias{monetPlot}
\title{Create monetPlots for your rmm results.}
\usage{
monetPlot(rmm, parameter, centrality = "median", lab = F, r = 3, sav = F)
}
\arguments{
\item{rmm}{A rmm object. rmm has to be run with monitor=T}

\item{parameter}{A string with the parameter name. The internal name has to be used, which are the rownames in the rmm reg.table output.}

\item{centrality}{A string specifying one of the following options: "median", "mean", "MAP", or "mode".}

\item{lab}{String to describe the parameter on the graph's x-axis. Optional. If not specified, the internal parameter name is used.}

\item{r}{Specify number of decimal places. Default equals 3.}

\item{sav}{TRUE or FALSE (default). If \code{TRUE}, the graph is saved to the current working directory as .png}
}
\value{
Returns a plot. The solid vertical is at 0 and the dashed vertical line is the mode of the posterior distributions.
}
\description{
The function \strong{monetPlot} creates a density plot of the posterior 
distribution of your model parameters and the traceplot that led to this density.
}
\examples{
data(coalgov)
m1 <- rmm(Surv(govdur, earlyterm) ~ 1 + mm(id(pid, gid), mmc(fdep), mmw(w ~ 1/n, constraint=1)) + majority + hm(id=cid, name=cname, type=RE, showFE=F),
          family="Weibull", monitor=T, data=coalgov)
monetPlot(m1, parameter="b.l1")

}
\author{
Benjamin Rosche <benjamin.rosche@gmail.com>
}
