% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my.rmse.R
\name{my.rmse}
\alias{my.rmse}
\title{my.rmse}
\usage{
my.rmse(prediction, response)
}
\arguments{
\item{prediction}{predicted value}

\item{response}{true value}
}
\value{
sqrt(mean((prediction-truth)^2))
}
\description{
root mean squared error
}

