% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right.R
\name{right}
\alias{right}
\title{right}
\usage{
right(text, num_char)
}
\arguments{
\item{text}{string}

\item{num_char}{number of chars from the right}
}
\value{
right substring
}
\description{
right of string.
}

