% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmm.R
\name{rmm}
\alias{rmm}
\title{Fit Bayesian micro-macro regression models using JAGS}
\usage{
rmm(formula, family = "Gaussian", priors, iter = 1000, burnin = 100,
  chains = 3, seed = NULL, data = NULL)
}
\arguments{
\item{formula}{A symbolic description of the model in form of an R formula. More details below.}

\item{family}{Character vector, either "Gaussian" or "Weibull"}

\item{priors}{A list with parameter or variable names as tags and their prior specification as values. More details below.}

\item{iter}{Total number of iterations}

\item{burnin}{Number of iterations that will be discarded}

\item{seed}{A random number}

\item{data}{The datasets must have level 1 as unit of analysis. More details below.}

\item{chain}{Number of chains}
}
\value{
JAGS output
}
\description{
Fit Bayesian micro-macro regression models using JAGS
}
\examples{
rmm(Y ~ 1 + X1 + mm(id(l1id, l2id), mmc(X2 + X3), mmw(w ~ 1/X4, constraint=1)),
    family="Gaussian",
    priors=list(var1="dnorm(0,0.0001)", tau.l1="dscaled.gamma(25, 1)"),
    data=data)

}
