% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nodedat}
\alias{nodedat}
\title{CILS4EU}
\format{
...
\describe{
  \item{Col1}{description of Col1}
...
}
}
\usage{
data(schoolnets)
}
\description{
...
schoolid school identifier
classid classroom identifier
youthid student identifier 
sex 
etn
parent_edu
parent_inc
test_lang
test_cogn
}
\examples{
data(schoolnets)
}
\keyword{datasets}
