% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep.cols.every.char.R
\name{sep.cols.every.char}
\alias{sep.cols.every.char}
\title{sep.cols.every.char}
\usage{
sep.cols.every.char(sc.data, x, maxchar)
}
\arguments{
\item{sc.data}{dataframe.}

\item{x}{column name in my.data to separate.}

\item{maxchar}{the maximum columns to split x into.}
}
\value{
return a dataframe with x split by ever char
}
\description{
Separate column in to one column for every character.
}

