% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep.cols.R
\name{sep.cols}
\alias{sep.cols}
\title{sep.cols}
\usage{
sep.cols(my.data, x)
}
\arguments{
\item{my.data}{dataframe}

\item{x}{column name in my.data to separate}
}
\value{
return a dataframe with x split by space
}
\description{
Separate column by space.
}

