% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToClip.R
\name{write.xclip}
\alias{write.xclip}
\title{write.xclip}
\usage{
write.xclip(x, selection = c("primary", "secondary", "clipboard"), ...)
}
\arguments{
\item{x}{object to put in clip}

\item{selection}{Default options "primary", "secondary", "clipboard"}

\item{...}{write table parameters}
}
\description{
Write to xClip (Linux)
}

