% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToClip.R
\name{writeclipl}
\alias{wcl}
\alias{writeclipl}
\title{writeclipl}
\usage{
writeclipl(x, ...)
}
\arguments{
\item{x}{object to put in clip}

\item{...}{write table parameters}
}
\description{
Write to Clipboard (Linux)
}

