% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToClip.R
\name{writeclipm}
\alias{wcm}
\alias{writeclipm}
\title{writeclipm}
\usage{
writeclipm(x, ...)
}
\arguments{
\item{x}{object to put in clip}

\item{...}{write table parameters}
}
\description{
Write to Clipboard (mac)
}

