% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToClip.R
\name{writeclipw}
\alias{wcw}
\alias{writeclipw}
\title{writeclipl}
\usage{
writeclipw()
}
\arguments{
\item{x}{object to put in clip}

\item{...}{write table parameters}
}
\description{
Write to Clipboard (windows)
}

